Cleanup of the code base #109
Don't freak out! I used S&R on all the files in the module to apply the code formatting we had all originally agreed on (spaces in brackets). I also optimized the imports,

This has absolutely no effect on functionality; this is merely a change in whitespace and getting rid of unused code.